refactor(app): delete unused useLaunchLegacyLPC
